Move duplicateview into browser module
